[Source] rename namespace from virtual to Abstracts
