+++ JDBSource/Interfaces/ITable.cs
@@ -1,4 +1,4 @@
-﻿using JDBSource.viruals;
+﻿using JDBSource.Abstracts;
 using System.Collections.Generic;
 using System.Threading.Tasks;
